refactor: extract techdocs auth to a new file
